begin porting... get basic formatting plumbing working
