+++ openvmm/openvmm_entry/src/cli_args.rs
@@ -626,10 +626,17 @@ pub enum DiskCliKind {
     Memory(u64),
     // memdiff:<kind>
     MemoryDiff(Box<DiskCliKind>),
-    // sql:<path>:<len>
-    Sqlite(PathBuf, u64),
-    // sqldiff:<path>:<kind>
-    SqliteDiff(PathBuf, Box<DiskCliKind>),
+    // sql:<path>[;create=<len>]
+    Sqlite {
+        path: PathBuf,
+        create_with_len: Option<u64>,
+    },
+    // sqldiff:<path>[;create]:<kind>
+    SqliteDiff {
+        path: PathBuf,
+        create: bool,
+        disk: Box<DiskCliKind>,
+    },
     // prwrap:<kind>
     PersistentReservationsWrapper(Box<DiskCliKind>),
     // file:<path>
@@ -669,13 +676,44 @@ impl FromStr for DiskCliKind {
             Some((kind, arg)) => match kind {
                 "mem" => DiskCliKind::Memory(parse_memory(arg)?),
                 "memdiff" => DiskCliKind::MemoryDiff(Box::new(arg.parse()?)),
-                "sql" => {
-                    let (path, len) = arg.split_once(':').context("expected path,len")?;
-                    DiskCliKind::Sqlite(path.into(), parse_memory(len)?)
-                }
+                "sql" => match arg.split_once(';') {
+                    Some((path, len)) => {
+                        let Some(len) = len.strip_prefix("create=") else {
+                            anyhow::bail!("invalid syntax after ';', expected 'create=<len>'")
+                        };
+
+                        DiskCliKind::Sqlite {
+                            path: path.into(),
+                            create_with_len: Some(parse_memory(len)?),
+                        }
+                    }
+                    None => DiskCliKind::Sqlite {
+                        path: arg.into(),
+                        create_with_len: None,
+                    },
+                },
                 "sqldiff" => {
-                    let (path, kind) = arg.split_once(':').context("expected path,kind")?;
-                    DiskCliKind::SqliteDiff(path.into(), Box::new(kind.parse()?))
+                    let (path_and_opts, kind) =
+                        arg.split_once(':').context("expected path,kind")?;
+                    let disk = Box::new(kind.parse()?);
+
+                    match path_and_opts.split_once(';') {
+                        Some((path, create)) => {
+                            if create != "create" {
+                                anyhow::bail!("invalid syntax after ';', expected 'create'")
+                            }
+                            DiskCliKind::SqliteDiff {
+                                path: path.into(),
+                                create: true,
+                                disk,
+                            }
+                        }
+                        None => DiskCliKind::SqliteDiff {
+                            path: path_and_opts.into(),
+                            create: false,
+                            disk,
+                        },
+                    }
                 }
                 "prwrap" => DiskCliKind::PersistentReservationsWrapper(Box::new(arg.parse()?)),
                 "file" => DiskCliKind::File(PathBuf::from(arg)),
