+++ vm/devices/storage/disklayer_sqlite/Cargo.toml
@@ -7,10 +7,11 @@ edition = "2021"
 rust-version.workspace = true
 
 [dependencies]
-disk_backend.workspace = true
 disk_backend_resources.workspace = true
-scsi_buffers.workspace = true
+disk_backend.workspace = true
+disk_layered.workspace = true
 guestmem.workspace = true
+scsi_buffers.workspace = true
 vm_resource.workspace = true
 
 inspect = { workspace = true, features = ["filepath"] }
