+++ vm/devices/storage/disklayer_sqlite/src/resolver.rs
@@ -1,72 +1,59 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-//! Resource resolver for SQLite-backed disk layers.
+//! Resource resolver for sqlite-backed disk layers.
 
-use crate::SqliteDisk;
-use async_trait::async_trait;
-use disk_backend::resolve::ResolveDiskParameters;
-use disk_backend::resolve::ResolvedDisk;
-use disk_backend_resources::SqliteDiffDiskHandle;
-use disk_backend_resources::SqliteDiskHandle;
-use std::path::Path;
-use vm_resource::declare_static_async_resolver;
-use vm_resource::kind::DiskHandleKind;
-use vm_resource::AsyncResolveResource;
-use vm_resource::ResourceResolver;
+use super::SqliteDiskLayer;
+use crate::FormatOnAttachSqliteDiskLayer;
+use disk_backend_resources::layer::SqliteDiskLayerFormatParams;
+use disk_backend_resources::layer::SqliteDiskLayerHandle;
+use disk_layered::resolve::ResolveDiskLayerParameters;
+use disk_layered::resolve::ResolvedDiskLayer;
+use vm_resource::declare_static_resolver;
+use vm_resource::kind::DiskLayerHandleKind;
+use vm_resource::ResolveResource;
 
-/// Resolver for a [`SqliteDiskHandle`] and [`SqliteDiffDiskHandle`].
-pub struct SqliteDiskResolver;
+/// Resolver for a [`SqliteDiskLayerHandle`].
+pub struct SqliteDiskLayerResolver;
 
-declare_static_async_resolver!(
-    SqliteDiskResolver,
-    (DiskHandleKind, SqliteDiskHandle),
-    (DiskHandleKind, SqliteDiffDiskHandle)
+declare_static_resolver!(
+    SqliteDiskLayerResolver,
+    (DiskLayerHandleKind, SqliteDiskLayerHandle)
 );
 
-#[async_trait]
-impl AsyncResolveResource<DiskHandleKind, SqliteDiskHandle> for SqliteDiskResolver {
-    type Output = ResolvedDisk;
+impl ResolveResource<DiskLayerHandleKind, SqliteDiskLayerHandle> for SqliteDiskLayerResolver {
+    type Output = ResolvedDiskLayer;
     type Error = anyhow::Error;
 
-    async fn resolve(
+    fn resolve(
         &self,
-        _resolver: &ResourceResolver,
-        rsrc: SqliteDiskHandle,
-        input: ResolveDiskParameters<'_>,
+        rsrc: SqliteDiskLayerHandle,
+        _input: ResolveDiskLayerParameters<'_>,
     ) -> Result<Self::Output, Self::Error> {
-        Ok(ResolvedDisk::new(SqliteDisk::new(
-            rsrc.len,
-            Path::new(&rsrc.dbhd_path),
-            input.read_only,
-        )?)?)
-    }
-}
+        let SqliteDiskLayerHandle {
+            dbhd_path,
+            format_dbhd,
+        } = rsrc;
 
-#[async_trait]
-impl AsyncResolveResource<DiskHandleKind, SqliteDiffDiskHandle> for SqliteDiskResolver {
-    type Output = ResolvedDisk;
-    type Error = anyhow::Error;
+        // TODO: not totally sure what the right input.read_only plumbing is, so
+        // I'm just ignoring it for now.
 
-    async fn resolve(
-        &self,
-        resolver: &ResourceResolver,
-        rsrc: SqliteDiffDiskHandle,
-        input: ResolveDiskParameters<'_>,
-    ) -> Result<Self::Output, Self::Error> {
-        let lower = resolver
-            .resolve(
-                rsrc.lower,
-                ResolveDiskParameters {
-                    read_only: true,
-                    _async_trait_workaround: &(),
+        let layer = if let Some(SqliteDiskLayerFormatParams {
+            logically_read_only,
+            len,
+        }) = format_dbhd
+        {
+            ResolvedDiskLayer::new(FormatOnAttachSqliteDiskLayer::new(
+                dbhd_path,
+                crate::IncompleteFormatParams {
+                    logically_read_only,
+                    len,
                 },
-            )
-            .await?;
-        Ok(ResolvedDisk::new(SqliteDisk::diff(
-            lower.0,
-            Path::new(&rsrc.dbhd_path),
-            input.read_only,
-        )?)?)
+            )?)
+        } else {
+            ResolvedDiskLayer::new(SqliteDiskLayer::new(dbhd_path, None)?)
+        };
+
+        Ok(layer)
     }
 }
